chore: separate getBaseSystemMessage logic from construct messages
